Brush up Makefile.unix.
In the beginning, Makefile.unix must have originated from Makefile.
After all those years, Makefile seems to have evolved, so there were
significant diversions between these two files.

With the help of a development branch 'makemake', both stes of evolution
can be nicely documented: (a) the similarity between Makefile.unix and
Makefile 3.64ad is made as clear as possible, (b) the changes between
the last state of Makefile.unix and its most recent incarnation is
clarified as well. Note that some GnuMake hackery is used to reduce the
volume of Makefile.unix, especially where the bulk of change files is
concerned.

Squashed commit of the following:

commit 095d5acf9eddec1a44a30c2db13c05dac640178b
Author: Andreas Scherer <andreas_scherer@alice.de>
Date:   Tue Sep 29 16:55:29 2015 +0200

    Update Makefile.unix and remove Makefile.

commit 16ad1f10c79c3b4e668712d3242bde42b00fddc4
Author: Andreas Scherer <andreas_scherer@alice.de>
Date:   Mon Sep 28 18:59:10 2015 +0200

    Makefile.unix originates from Makefile.

commit efae0d99a95ab9d52b30b90a9dd10fbc8fed5de7
Author: Andreas Scherer <andreas_scherer@alice.de>
Date:   Mon Sep 28 16:07:23 2015 +0200

    Makefile from cweb 3.64ad.