Compactify two-part %setup.
Using the default action allows to combine the two %setup lines. In this
case, '-a 1' may be read as 'add Source1:'. ('-c' creates the build area
and cd's into it, so here '-a' and '-b' would work in similar ways, as
long as both source drops do not extract to a top-level directory.)