Use '%autosetup' to set up Git in BUILD.
There are tons of "change files" in CWEBBIN but no "patch files" (yet).
Anyway, it may be useful to have the BUILD arena set up with SCM
support. For me this means "Git", however, others are supported as well.
After the '%prep'aration step, full versioning support is provided.

Just for fun I add '-v' to see the extraction and initial commit.