Revert "Add file suffix for xdvipdfmx."
This reverts commit b19a1fa0acc73fbf0be7b10b7bae1f2c4b1cb473.

At least the Build tree of TeXlive (2018, 2019dev) has the necessary fix
for this issue. As I really don't like to pick such upstream nits in
cwebbin, I revert to the original pdf creation programme.

If you really absolutely need to create fullmanual, uncomment the pdftex
line in Makefile.unix and all will be fine. Or wait for TeXlive 2019.