Add TeXLive's changes last in line.
The main idea is that TL's CWEB shall be based on CWEBbin. It seems to
be feasible to merge all of our stuff with the original CWEB sources and
apply the TL changes to those either included here or as single add-ons
in the TL build ecosystem. Otherwise the whole build mechanism would
have to be analyzed, disected, and amended (using 'tie' or 'ctie').