Avoid naming conflict for headerfile.
'cweb.h' is a much too generic name for a local headerfile. After many
years it was controlable in the CWEBbin environment, but with the
current attempt to integrate CB with the TeXLive ecosystem, it clashes
with TL's 'cweb.h', which is used for a completely different purpose.

Instead of loading the unadorned 'cweb.h' from the $(CATINCLUDE) search
path, its better to expressly request the relative include path.