Automatic integration in TeXLive.
Merge all our changefiles with the interface changefile *-texlive.ch
into a set of *-w2c.ch files required by TeXLive's build ecosystem as
drop-in replacements for the files with the same names.

TODO: Create a stand-alone source package (pax/tar) for publication.
No %build is necessary, it will be done in texlive-source. Simply
collect the set of files for distribution.