Demote WMERGE to "example only".
TeX Live does _not_ provide a 'wmerge' executable, and in light of
'tie' and 'ctie' its capabilities are not really unique.

If you really absolutely need to merge a single (!) change file into an
input CWEB file, resolving all '@i'nclude files on the way, use 'ctie':

$ wmerge input.w change.ch output.w

is virtually the same as

$ ctie -m output.w input.w change.ch

but the latter can apply much more than just 'change.ch', like

$ ctie -m output.w input.w change1.ch chamge2.ch change3.ch ...

LATE BREAKING NEWS: Just by opening ../cweb/examples/wmerge.w to look up
the invocation rules, I introduced a "reading error" into the source module.
The double "and and" is immediately fixed to "and an" with a patch. Sorry!