Add type declarations for system headers.
See page 232 of "Digital Typography" for details.

This is complete for "ctwill.w", but by far not a comprehensive
description of all external elements in the C standard library.

The general recipe for twilled and TeXed output is:

$> CWEBINPUTS=cwebinputs: ctangle ctwill
$> g++ -o ctwill ctwill.c common.o
$> CWEBINPUTS=cwebinputs ./ctwill ctwill
$> CWEBINPUTS=cwebinputs ./ctwill ctwill
$> pdftex ctwill.tex
$> pdftex ctwill.tex
$> refsort < ctwill.ref > ctwill.sref
$> pdftex ctwill.tex

TODO: Write an additional 'ctwill.bux' with the remaining elements from
'common.w'.