Integrate CTWILL in TeX Live.
CTWILL is the 'evil twin' of CWEAVE, so most of the code is reused.
The necessary changes that mogrify 'cweave.w' into 'ctwill.w' come
from the 'local' branch of the CTWILL project on Github:
https://github.com/ascherer/ctwill/tree/local

The original tarball for CTWILL should be available from
ftp://ftp.cs.stanford.edu/pub/ctwill, but this link is broken today.

The *.diffs files from DEK were transformed into genuine change files in
the 'local' branch above and are applied to cweave.w and prod.w here.

The two utility programs 'refsort.w' and 'twinx.w' were improved for
better C/C++-compilability also in the CWEBBIN project.

The original macros 'ctwimac.tex' and 'proofmac.tex' and their German
translation wrappers are added as is.

Using CTWILL is somewhat involved, but the procedure boils down to
$> ctwill [+P] CWEBSOURCE CHANGEFILE # prime the pump
$> ctwill [+P] CWEBSOURCE CHANGEFILE # prepare TeX input
$> tex CTWILLOUTPUT # prime the pump
$> refsort < CTWILLOUTPUT.ref > CTWILLOUTPUT.sref
$> tex CTWILLOUTPUT # create the final product

Option '+P' for CTWILL replaces 'ctwimac.tex' by 'proofmac.tex' in the
first line of CTWILLOUTPUT for 'proof reading' the index entries per
section. This form is easier to read than the actual 'mini-indexes'.