Use individual table-of-contents file.
Standard WEB uses generic 'CONTENTS.tex' as filename for the
table-of-contents information.  Running multiple WEB codes in
a batch run -- like with the 'makeall' script -- leads to conflicts,
because the _first_ run of the $n+1$st WEB code picks up the toc
file from the _second_ run of the $n$th WEB code, and in most cases
there are irresolvable references.  To avoid these collisions, we
use the naming scheme from CWEB.