Fake solution for issue #32, headline after TOC.
There's a timing problem with all the to-and-fro'ing between '\con' and
the first '\N'; somehow the '\mark' in '\MN' gets lost and the defaults
'\nullsec' and '\.{CWEB} output' show up.  These values most likely will
break any other programs than CWEB itself.

I guess there's a solution in relation to the various '\eject' calls.
The mark set by '\MN' (inside the first '\N') must appear _before_ the
'\eject' of the table-of-contents page, because the '\topmark' will be
used for the headline of the first page _after_ the TOC.

For our purposes, it's perfectly fine to use fixed values, both for WEB
programs (that's easy; _every_ program starts with 'section 1') and CWEB
programs (all CWEB modules start with their 'Introduction').