Accommodate Perl v5.8.4 on Solaris.
Somehow, the double check (in ctwill.test and in proofsort) for the
availability of module 'Unicode::Collate' succeeds, but the invocation
of creator 'new()' fails.

We try to fix this issue by including both statements in the 'eval'
group, and hope for the best.

(Also, '$progname' must be initialized in the 'BEGIN' group.)